Convert admin_level to int so it can be filtered better.

+++ sql/admin.sql
@@ -1,4 +1,4 @@
-select name, admin_level, way as geom
+select name, admin_level::int, way as geom
 FROM planet_osm_line
 WHERE boundary = 'administrative'
   AND admin_level IN ('2','3','4')
